+++ R/divergence_minimizers.R
@@ -78,7 +78,18 @@ divmin <- function(
     )
   }
 
-  if (length(formulas) < getOption("projpred.prll_prj_trigger", Inf)) {
+  if (is.null(getOption("projpred.parallel_proj_trigger")) &&
+      !is.null(getOption("projpred.prll_prj_trigger"))) {
+    warning(
+      "Global option `projpred.prll_prj_trigger` is deprecated. Please use ",
+      "global option `projpred.parallel_proj_trigger` instead. Now using the ",
+      "value from global option `projpred.prll_prj_trigger`."
+    )
+    prll_prj_trigger <- getOption("projpred.prll_prj_trigger")
+  } else {
+    prll_prj_trigger <- getOption("projpred.parallel_proj_trigger", Inf)
+  }
+  if (length(formulas) < prll_prj_trigger) {
     # Sequential case. Actually, we could simply use ``%do_projpred%` <-
     # foreach::`%do%`` here and then proceed as in the parallel case, but that
     # would require adding more "hard" dependencies (because packages 'foreach'
@@ -660,7 +671,18 @@ divmin_augdat <- function(
   projpred_ws_aug <- unclass(projpred_ws_aug)
   attr(projpred_ws_aug, "ndiscrete") <- NULL
 
-  if (ncol(projpred_ws_aug) < getOption("projpred.prll_prj_trigger", Inf)) {
+  if (is.null(getOption("projpred.parallel_proj_trigger")) &&
+      !is.null(getOption("projpred.prll_prj_trigger"))) {
+    warning(
+      "Global option `projpred.prll_prj_trigger` is deprecated. Please use ",
+      "global option `projpred.parallel_proj_trigger` instead. Now using the ",
+      "value from global option `projpred.prll_prj_trigger`."
+    )
+    prll_prj_trigger <- getOption("projpred.prll_prj_trigger")
+  } else {
+    prll_prj_trigger <- getOption("projpred.parallel_proj_trigger", Inf)
+  }
+  if (ncol(projpred_ws_aug) < prll_prj_trigger) {
     # Sequential case. Actually, we could simply use ``%do_projpred%` <-
     # foreach::`%do%`` here and then proceed as in the parallel case, but that
     # would require adding more "hard" dependencies (because packages 'foreach'